\name{RcmdrPlugin.doBy-package}
\alias{RcmdrPlugin.doBy-package}
\alias{RcmdrPlugin.doBy}
\docType{package}
\title{
Rcmdr doBy Plug-In
}
\description{
This package provides an R-commander plug-in for the core doBy functions in the doBy package, which helps with computing groupwise summary statistics (similar to PROC SUMMARY in the SAS system).
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.doBy\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2011-12-12\cr
License: \tab GPL 2\cr
}

}
\author{
Jonathan Lee <jlee253@uwo.ca>

Maintainer: Jonathan Lee <jlee253@uwo.ca>
}
\references{
Fox, J. (2007) Extending the R Commander by ``plug in'' packages. \emph{R News}, 7(3): 46-52.
}

\keyword{ package }
\seealso{
  \code{\link{Rcmdr}}
  \code{\link{doBy}}
}
