\name{summaryByGUI}
\alias{summaryByGUI}
\alias{orderByGUI}
\alias{splitByGUI}
\alias{sampleByGUI}
\title{
Functions to trigger the GUI menu calls in Rcmdr
}
\description{
These functions are called by Rcmdr menu items and are not meant to be called by the end user.
}
\usage{
summaryByGUI()
orderByGUI()
splitByGUI()
sampleByGUI()
}
%- maybe also 'usage' for other objects documented here.
\details{
}
\value{
Returns \code{FALSE} upon window close.
}
\author{
Jonathan Lee <jlee253@uwo.ca>
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ device }
